+++ packages/http-utils/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@libp2p/http-utils-v1.0.1](https://github.com/libp2p/js-libp2p-http/compare/@libp2p/http-utils-1.0.0...@libp2p/http-utils-1.0.1) (2025-06-27)
+
+### Bug Fixes
+
+* update multiaddr methods ([#117](https://github.com/libp2p/js-libp2p-http/issues/117)) ([e683721](https://github.com/libp2p/js-libp2p-http/commit/e683721c2025fe04bd1145fd60082b5575c4c9ac))
+
 ## @libp2p/http-utils-v1.0.0 (2025-05-28)
 
 ### ⚠ BREAKING CHANGES